Version 1.02
Found that Cream had an abbreviation file, so I took any words that were in there that I didn't already have and added them (about thirty, with some day of the week capitalization.)  Changed output of generator to recognize when a fix was just a capitalization.  Reduced size of final file by changing from 'iabbrev' to 'ia' (8000*5 characters!)  Also went through and deleted corrections of dubious quality and fixed some problems in the source file.  Added a couple of things that I've made mistakes on in the past two months.  All in all, a bunch of nice changes, less lines and size and more autocorrectionness.