Version 1.03
Thanks to Gilles Bertrand for sending me a list containing a bunch of new corrections!  Many of the corrections are longer or more technical in nature.  In addition, I created a file in the GitHub repository that contains multiple word corrections in case this is ever easy to do in the future with Vim.  I also added a README in the tar file itself.